+++ readthedocs/proxito/views/hosting.py
@@ -16,6 +16,7 @@ from readthedocs.api.v2.permissions import IsAuthorizedToViewVersion
 from readthedocs.api.v3.serializers import (
     BuildSerializer,
     ProjectSerializer,
+    RelatedProjectSerializer,
     VersionSerializer,
 )
 from readthedocs.builds.constants import BUILD_STATE_FINISHED, LATEST
@@ -245,7 +246,13 @@ class NoLinksMixin:
 # on El Proxito.
 #
 # See https://github.com/readthedocs/readthedocs-ops/issues/1323
+class RelatedProjectSerializerNoLinks(NoLinksMixin, RelatedProjectSerializer):
+    pass
+
+
 class ProjectSerializerNoLinks(NoLinksMixin, ProjectSerializer):
+    related_project_serializer = RelatedProjectSerializerNoLinks
+
     def __init__(self, *args, **kwargs):
         resolver = kwargs.pop("resolver", Resolver())
         super().__init__(
@@ -445,6 +452,9 @@ class AddonsResponseBase:
                 "analytics": {
                     "code": settings.GLOBAL_ANALYTICS_CODE,
                 },
+                "resolver": {
+                    "filename": filename,
+                },
             },
             # TODO: the ``features`` is not polished and we expect to change drastically.
             # Mainly, all the fields including a Project, Version or Build will use the exact same
@@ -486,6 +496,7 @@ class AddonsResponseBase:
                     #     "branch": version.identifier if version else None,
                     #     "filepath": "/docs/index.rst",
                     # },
+                    "position": project.addons.flyout_position,
                 },
                 "customscript": {
                     "enabled": project.addons.customscript_enabled,
@@ -628,7 +639,7 @@ class AddonsResponseBase:
         This response is only enabled for external versions,
         we do the comparison between the current version and the latest version.
         """
-        if not version.is_external:
+        if not version.is_external and not settings.RTD_FILETREEDIFF_ALL:
             return None
 
         if not project.addons.filetreediff_enabled: